add test utility functions module
lint
